fix: move styles to separate files
