--- src/components/common/ContentWrapper.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const ContentWrapper = styled.div`
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  height: 100%;
-  background-color: #f7f9ff;
-  width: 100%;
-  overflow-x: auto;
-  padding: 20px;
-  font-size: 20px;
-`;
