+++ src/components/common/Form.tsx
@@ -1,23 +1,4 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled.form`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  max-width: 300px;
-  width: 100%;
-
-  & button {
-    margin-top: 30px;
-    width: 95px;
-  }
-
-  & fieldset {
-    width: 100%;
-    margin: 15px 0 15px;
-  }
-`;
+import { Wrapper } from './CommonStyles';
 
 type Props = {
   children: React.ReactElement | React.ReactElement[];
