--- src/components/common/PageWrapper.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const PageWrapper = styled.div`
-  display: flex;
-  height: 100vh;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-`;
