+++ src/components/common/index.ts
@@ -2,5 +2,3 @@ export { default as SubmitButton } from './Button';
 export { default as NavigationButton } from './NavigationButton';
 export { default as Form } from './Form';
 export { default as Input } from './Input';
-export { ContentWrapper } from './ContentWrapper';
-export { PageWrapper } from './PageWrapper';
