+++ src/pages/RegistrationPage.tsx
@@ -1,83 +1,11 @@
 import { Header, Footer } from '../components';
-import { PageWrapper } from '../components/common';
 import { useForm, SubmitHandler } from 'react-hook-form';
-import ShowButton from '../components/common/ShowButton';
-import styled from 'styled-components';
+import ShowButton from '../components/common/SwitchButton';
 import { useState } from 'react';
 import { CustomerDraft } from '@commercetools/platform-sdk';
 import { signInCustomer } from '../services/Client';
 import { useNavigate } from 'react-router-dom';
-
-const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  max-width: 300px;
-  width: 100%;
-
-  & button {
-    margin-top: 30px;
-    width: 95px;
-  }
-
-  & fieldset {
-    width: 100%;
-    margin: 15px 0 15px;
-  }
-`;
-const InputElem = styled.input`
-  background-color: #fff;
-  border-radius: 3px;
-  border: 1px solid #7aa7c7;
-  color: #39739d;
-  font-size: 16px;
-  padding: 10px;
-  margin-bottom: 10px;
-  width: 100%;
-
-  &:focus {
-    outline: none;
-    border: 2px solid #39739d;
-  }
-`;
-
-const ButtonSubmit = styled.button`
-  background-color: #ada5f9;
-  border-radius: 3px;
-  border: 1px solid #ada5f9;
-  color: #511f31;
-  cursor: pointer;
-  padding: 10px 10px;
-  font-size: 18px;
-
-  &:hover,
-  :focus {
-    background-color: #f7f9ff;
-    color: #2c5777;
-  }
-  &:disabled {
-    background-color: #c0c0c0;
-    color: black;
-  }
-`;
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  background-color: #f7f9ff;
-  width: 100%;
-
-  padding: 20px;
-  font-size: 20px;
-`;
-
-const ErrorsText = styled.p`
-  color: red;
-  font-size: 13px;
-`;
+import { ButtonSubmit, Container, ErrorsText, Form, InputElem, PageWrapper } from './PagesStyles';
 
 export default function RegistrationPage() {
   const headerButtons = [
